re-adjusting selector and modifier wrappers
moved into corresponding modules instead of keeping them in separate files
