+++ src/compas_rhino/modifiers/__init__.py
@@ -93,7 +93,5 @@ from .edgemodifier import *
 from .facemodifier import *
 
 from .modifier import *
-from .meshmodifier import *
-from .networkmodifier import *
 
 __all__ = [name for name in dir() if not name.startswith('_')]
