+++ src/compas_rhino/modifiers/vertexmodifier.py
@@ -33,7 +33,16 @@ else:
         compas.raise_if_ironpython()
 
 
-__all__ = ['VertexModifier']
+__all__ = [
+    'VertexModifier',
+
+    'mesh_update_vertex_attributes',
+    'mesh_move_vertex',
+    'mesh_move_vertices',
+
+    'network_move_vertex',
+    'network_update_vertex_attributes'
+    ]
 
 
 def rhino_update_named_values(names, values, message='', title='Update named values'):
@@ -186,6 +195,120 @@ class VertexModifier(object):
                             self.vertex[key][name] = value
             return True
         return False
+
+
+def mesh_update_vertex_attributes(mesh, keys, names=None):
+    """Update the attributes of the vertices of a mesh.
+
+    Parameters
+    ----------
+    mesh : compas.datastructures.Mesh
+        A mesh object.
+    keys : tuple, list
+        The keys of the vertices to update.
+    names : tuple, list (None)
+        The names of the atrtibutes to update.
+        Default is to update all attributes.
+
+    Returns
+    -------
+    bool
+        ``True`` if the update was successful.
+        ``False`` otherwise.
+
+    See Also
+    --------
+    * :func:`mesh_update_attributes`
+    * :func:`mesh_update_edge_attributes`
+    * :func:`mesh_update_face_attributes`
+
+    """
+    return VertexModifier.update_vertex_attributes(mesh, keys, names=names)
+
+
+def mesh_move_vertex(mesh, key, constraint=None, allow_off=False):
+    """Move on vertex of the mesh.
+
+    Parameters
+    ----------
+    mesh : compas.datastructures.Mesh
+        A mesh object.
+    key : str
+        The vertex to move.
+    constraint : Rhino.Geometry (None)
+        A Rhino geometry object to constrain the movement to.
+        By default the movement is unconstrained.
+    allow_off : bool (False)
+        Allow the vertex to move off the constraint.
+
+    """
+    return VertexModifier.move_vertex(mesh, key, constraint=constraint, allow_off=allow_off)
+
+
+def mesh_move_vertices(mesh, keys):
+    """Move on vertices of the mesh.
+
+    Parameters
+    ----------
+    mesh : compas.datastructures.Mesh
+        A mesh object.
+    keys : list
+        The vertices to move.
+    constraint : Rhino.Geometry (None)
+        A Rhino geometry object to constrain the movement to.
+        By default the movement is unconstrained.
+    allow_off : bool (False)
+        Allow the vertex to move off the constraint.
+
+    """
+    return VertexModifier.move_vertices(mesh, keys)
+
+
+def network_move_vertex(network, key, constraint=None, allow_off=False):
+    """Move on vertex of the network.
+
+    Parameters
+    ----------
+    network : compas.datastructures.Network
+        A network object.
+    key : str
+        The vertex to move.
+    constraint : Rhino.Geometry (None)
+        A Rhino geometry object to constrain the movement to.
+        By default the movement is unconstrained.
+    allow_off : bool (False)
+        Allow the vertex to move off the constraint.
+
+    """
+    return VertexModifier.move_vertex(network, key, constraint=constraint, allow_off=allow_off)
+
+
+def network_update_vertex_attributes(network, keys, names=None):
+    """Update the attributes of the vertices of a network.
+
+    Parameters
+    ----------
+    network : compas.datastructures.Network
+        A network object.
+    keys : tuple, list
+        The keys of the vertices to update.
+    names : tuple, list (None)
+        The names of the atrtibutes to update.
+        Default is to update all attributes.
+
+    Returns
+    -------
+    bool
+        ``True`` if the update was successful.
+        ``False`` otherwise.
+
+    See Also
+    --------
+    * :func:`network_update_attributes`
+    * :func:`network_update_edge_attributes`
+
+    """
+    return VertexModifier.update_vertex_attributes(network, keys, names=names)
 
 
 # ==============================================================================
